+++ lib/src/features/users/presentation/pages/users_page.dart
@@ -2,12 +2,14 @@ import 'package:bflow_client/src/core/config/config.dart';
 import 'package:bflow_client/src/core/constants/colors.dart';
 import 'package:bflow_client/src/core/extensions/build_context_extensions.dart';
 import 'package:bflow_client/src/core/widgets/action_button_widget.dart';
+import 'package:bflow_client/src/core/widgets/confirmation_widget.dart';
 import 'package:bflow_client/src/core/widgets/failure_widget.dart';
 import 'package:bflow_client/src/core/widgets/page_container_widget.dart';
 import 'package:bflow_client/src/features/users/domain/entities/user_entity.dart';
 import 'package:bflow_client/src/features/users/presentation/bloc/users_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:go_router/go_router.dart';
 
 import '../widgets/write_user_widget.dart';
 
@@ -141,9 +143,22 @@ class UsersPage extends StatelessWidget {
               tooltip: 'Edit',
             ),
             IconButton(
-              onPressed: () => context
-                  .read<UsersBloc>()
-                  .add(DeleteUserEvent(userId: user.id!)),
+              onPressed: () {
+                context.showCustomModal(
+                  ConfirmationWidget(
+                    title: "Delete user",
+                    description:
+                        "Are you sure you want to delete user \"${user.username}\"?",
+                    onConfirm: () {
+                      context
+                          .read<UsersBloc>()
+                          .add(DeleteUserEvent(userId: user.id!));
+                      context.pop();
+                    },
+                    confirmText: "Delete",
+                  ),
+                );
+              },
               color: AppColor.blue,
               icon: const Icon(
                 Icons.delete_outline_outlined,
